update logging snippets, clean up conftest
